requirements.txt
joblib==1.3.2
Django==5.0.1
djangorestframework==3.14.0
scikit-learn==1.4.0
django-cors-headers==4.2.0